refactor: move vaa parse and verificationt to k
