+++ keeper/keeper.go
@@ -23,6 +23,8 @@ package keeper
 import (
 	"context"
 	"encoding/json"
+	"fmt"
+	"strconv"
 
 	"cosmossdk.io/collections"
 	collcodec "cosmossdk.io/collections/codec"
@@ -36,6 +38,8 @@ import (
 	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
 	clienttypes "github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
 	"github.com/cosmos/ibc-go/v8/modules/core/24-host"
+	"github.com/ethereum/go-ethereum/common"
+	vaautils "github.com/wormhole-foundation/wormhole/sdk/vaa"
 
 	"github.com/noble-assets/wormhole/types"
 )
@@ -156,4 +160,41 @@ func (k *Keeper) PostMessage(ctx context.Context, signer string, message []byte,
 	)
 
 	return err
+}
+
+func (k *Keeper) ParseAndVerifyVAA(ctx context.Context, bz []byte) (*vaautils.VAA, error) {
+	vaa, err := vaautils.Unmarshal(bz)
+	if err != nil {
+		return nil, errors.Wrapf(types.ErrInvalidVAA, "failed to unmarshal: %v", err)
+	}
+
+	hash := vaa.SigningDigest().Bytes()
+	if has, err := k.VAAArchive.Has(ctx, hash); err != nil || has {
+		return nil, types.ErrAlreadyExecutedVAA
+	}
+
+	guardianSet, err := k.GuardianSets.Get(ctx, vaa.GuardianSetIndex)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get guardian set %d from state", vaa.GuardianSetIndex)
+	}
+
+	blockTime := uint64(k.headerService.GetHeaderInfo(ctx).Time.Unix())
+	// TODO: is zero a no expiration?
+	if guardianSet.ExpirationTime != 0 && guardianSet.ExpirationTime < blockTime {
+		return nil, fmt.Errorf("guardian set %d is expired", vaa.GuardianSetIndex)
+	}
+
+	var addresses []common.Address
+	for _, address := range guardianSet.Addresses {
+		addresses = append(addresses, common.BytesToAddress(address))
+	}
+	if err := vaa.Verify(addresses); err != nil {
+		return nil, errors.Wrap(err, "failed to verify vaa")
+	}
+
+	if err := k.VAAArchive.Set(ctx, hash, collections.Join(vaa.MessageID(), true)); err != nil {
+		return nil, errors.Wrap(err, "failed to set vaa in state")
+	}
+
+	return vaa, nil
 }
