+++ keeper/wormhole_core.go
@@ -25,10 +25,8 @@ import (
 	"fmt"
 	"strconv"
 
-	"cosmossdk.io/collections"
 	"cosmossdk.io/core/event"
 	"cosmossdk.io/errors"
-	"github.com/ethereum/go-ethereum/common"
 	vaautils "github.com/wormhole-foundation/wormhole/sdk/vaa"
 
 	"github.com/noble-assets/wormhole/types"
@@ -37,43 +35,6 @@ import (
 // GuardianSetExpiry defines how long a guardian set should remain active for
 // after being replaced, before then expiring. Currently, 24 hours.
 const GuardianSetExpiry = 86400
-
-func (k *Keeper) ParseAndVerifyVAA(ctx context.Context, bz []byte) (*vaautils.VAA, error) {
-	vaa, err := vaautils.Unmarshal(bz)
-	if err != nil {
-		return nil, errors.Wrapf(types.ErrInvalidVAA, "failed to unmarshal: %v", err)
-	}
-
-	hash := vaa.SigningDigest().Bytes()
-	if has, err := k.VAAArchive.Has(ctx, hash); err != nil || has {
-		return nil, types.ErrAlreadyExecutedVAA
-	}
-
-	guardianSet, err := k.GuardianSets.Get(ctx, vaa.GuardianSetIndex)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get guardian set %d from state", vaa.GuardianSetIndex)
-	}
-
-	blockTime := uint64(k.headerService.GetHeaderInfo(ctx).Time.Unix())
-	// TODO: is zero a no expiration?
-	if guardianSet.ExpirationTime != 0 && guardianSet.ExpirationTime < blockTime {
-		return nil, fmt.Errorf("guardian set %d is expired", vaa.GuardianSetIndex)
-	}
-
-	var addresses []common.Address
-	for _, address := range guardianSet.Addresses {
-		addresses = append(addresses, common.BytesToAddress(address))
-	}
-	if err := vaa.Verify(addresses); err != nil {
-		return nil, errors.Wrap(err, "failed to verify vaa")
-	}
-
-	if err := k.VAAArchive.Set(ctx, hash, collections.Join(vaa.MessageID(), true)); err != nil {
-		return nil, errors.Wrap(err, "failed to set vaa in state")
-	}
-
-	return vaa, nil
-}
 
 func (k *Keeper) HandleCoreGovernancePacket(ctx context.Context, pkt types.GovernancePacket) error {
 	switch pkt.Action {
